+++ lib/view/widgets/display_nodes/gen/autocomplete.g.dart
@@ -1,0 +1,8 @@
+/// ===================================================
+/// Power By 张风捷特烈 --- Generated file. Do not edit.
+/// github: https://github.com/toly1994328
+/// ===================================================
+
+part of 'node.g.dart';
+
+Map<String, dynamic> get _autocompleteData => {"AutocompleteDemo1":{"title":"TolyAutocomplete 基本使用","desc":"【optionsBuilder】 : 选项构造器   【AutocompleteOptionsBuilder<T>】\n【fieldViewBuilder】 : 选择时回调 【AutocompleteFieldViewBuilder】\n【onSelected】 : 选择时回调   【ValueChanged<T>】","code":"assets/code_res/autocomplete_demo1.txt"},"AutocompleteDemo2":{"title":"TolyAutocomplete 泛型使用","desc":"使用 itemBuilder 可根据泛型自定义浮层条目构建、","code":"assets/code_res/autocomplete_demo2.txt"}};