+++ lib/view/widgets/form/form.dart
@@ -1,3 +1,4 @@
 export 'input/input.dart';
 export 'select/select.dart';
 export 'checkbox/checkbox.dart';
+export 'autocomplete/autocomplete.dart';