+++ modules/tolyui/CHANGELOG.md
@@ -1,11 +1,15 @@
 ## 0.0.1
 
-* +widget: TolyLink 
+* [⭐️+widget⭐️]: TolyLink 
 
 ## 0.0.2
 
 * [tolyui_feedback](https://pub.dev/packages/tolyui_feedback)
 * [tolyui_message](https://pub.dev/packages/tolyui_message)
 
-* +20: +widget: TolyCollapse 
+* +20: [⭐️+widget⭐️]: TolyCollapse 
 * +21: tolyui_navigation: 0.0.8+10
+
+## 0.0.3
+* +1: [⭐️+widget⭐️]: TolyAutocomplete 
+